gateway/requirements.txt
flask==3.0.0
requests==2.31.0
pillow==10.1.0
numpy==1.26.2
